polish test_gpt
