+++ tests/components_to_test/__init__.py
@@ -1,1 +1,1 @@
-from . import repeated_computed_layer, resnet, nested_model, bert, no_leaf_module, simple_net
+from . import repeated_computed_layer, resnet, nested_model, bert, no_leaf_module, simple_net, gpt
